% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geminiEDA.R
\name{eda_app}
\alias{eda_app}
\title{Run the Shiny app}
\usage{
eda_app(data = NULL)
}
\arguments{
\item{data}{Optional dataframe to use in the Shiny app.}
}
\description{
This function launches the Shiny app for generating EDA reports.
It can accept a dataframe from the current environment.
}
