% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_functions.R
\name{eda_biv_autoplot}
\alias{eda_biv_autoplot}
\title{Create a univariate plot based on variable type}
\usage{
eda_biv_autoplot(
  data,
  covar,
  out,
  input_type_cov = NULL,
  input_type_out = NULL
)
}
\arguments{
\item{data}{The dataset containing the variable.}

\item{var}{The name of the variable to plot.}
}
\value{
A ggplot2 object displaying the histogram or bar plot based on the variable type.
}
\description{
Generates a univariate plot for a specified variable in the dataset, distinguishing between numeric and categorical variables.
}
