% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_functions.R
\name{eda_biv_summary}
\alias{eda_biv_summary}
\title{Create a bivariate summary table based on variable types}
\usage{
eda_biv_summary(data, covar, out)
}
\arguments{
\item{data}{The dataset containing the variables.}

\item{covar}{The name of the covariate variable.}

\item{out}{The name of the outcome variable.}
}
\value{
A gtsummary table object with statistics based on the types of the covariate and outcome variables.
}
\description{
Generates a summary table comparing two variables in the dataset based on their types (numeric or categorical).
}
