% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_functions.R
\name{generate_revealjs_presentation}
\alias{generate_revealjs_presentation}
\title{Generate a Quarto Reveal.js Presentation Document}
\usage{
generate_revealjs_presentation(file_path)
}
\arguments{
\item{file_path}{Character. The file path to the EDA report (in .qmd format) on which the presentation will be based.}
}
\value{
NULL. This function does not return a value; instead, it creates a new Quarto presentation document as a side effect.
}
\description{
This function generates a Quarto presentation document in the Reveal.js format, based on an EDA (Exploratory Data Analysis) report created by a Shiny app. The resulting presentation is ready for use and can be further edited if needed.
}
\details{
The function reads the specified EDA report, extracts relevant sections and metadata (title, author, date), and compiles them into a new Quarto document with a Reveal.js format. The generated presentation file is named `presentation_<basename>.qmd`, where `<basename>` is the original file name without its extension.
}
\examples{
\dontrun{
generate_revealjs_presentation("path_to_file.qmd")
}

}
