% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_eda_markdown.R
\name{populate_script_download}
\alias{populate_script_download}
\title{Generate R Markdown Script for EDA Report}
\usage{
populate_script_download(
  filename,
  covariates,
  outcomes,
  author,
  project,
  data_name
)
}
\arguments{
\item{filename}{Character. The file path to save the R Markdown script.}

\item{covariates}{Character vector. The selected covariates for the EDA.}

\item{outcomes}{Character vector. The selected outcomes for the EDA.}

\item{author}{Character. The author of the report.}

\item{project}{Character. The project name.}

\item{data_name}{Character. The name of the dataframe to be used in the EDA report.}
}
\value{
NULL. This function is used for its side effects of writing to a file.
}
\description{
This function creates an R Markdown script for an Exploratory Data Analysis (EDA) report.
The script includes univariate and bivariate analyses based on selected covariates and outcomes.
}
\examples{
\dontrun{
populate_script_download("report.Rmd", covariates = c("age", "gender"),
                         outcomes = c("income"), author = "John Doe",
                         project = "Sample Project", data_name = "my_data")
}
}
