% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_functions.R
\name{scale_colour_GEMINI}
\alias{scale_colour_GEMINI}
\title{Set a custom color scale for ggplot2 plots in the GEMINI style}
\usage{
scale_colour_GEMINI(...)
}
\arguments{
\item{...}{Additional parameters passed to `discrete_scale()`.}
}
\value{
A ggplot2 scale object for color aesthetics.
}
\description{
This function defines a custom color scale for ggplot2 plots according to the GEMINI style, using a predefined set of colors.
}
