% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_functions.R
\name{theme_GEMINI}
\alias{theme_GEMINI}
\title{Set a custom GEMINI theme for ggplot2 plots}
\usage{
theme_GEMINI(base_size = 14, base_family = "helvetica")
}
\arguments{
\item{base_size}{Base font size for the plot elements (default is 14).}

\item{base_family}{Base font family for the plot text (default is "helvetica").}
}
\value{
A ggplot2 theme object that can be added to ggplot plots.
}
\description{
This function applies a custom GEMINI theme to ggplot2 plots, adjusting elements like text, background, axes, and legends to fit a predefined style.
}
